Moved test related to timeout into another class to circumvent FileLengthCheck
